Translation Update and Revert Mod Name
Created translation folder for v1.5 and updated as appropriate for the mod in its current version. This new folder should serve as the base for all translation mods in the future. Rimworld has rich support for translation via separate mods, and this should be used in the future. Removed outdated high level translations to reflect this policy.
Previously, I had changed the about.xml mod name to match the Steam Workshop title, but this breaks any mod that would try and patch mine. There are a lot of mods that do this. Expecting all other mods to update to this new name would be silly, so I am reverting this change. Users still subscribed to the original R++ will just have to look a little closer before activating the mod.