Update RPP_VanillaFishingExpanded_Compatibility.xml
That's what I get for declaring success without actually testing with/without the new DLC. VFE cedes the fishing worktype to Odyssey, so a patch is only required when running VFE without Odyssey.
Side note: this means robots will be able to fish with the mod, but not the DLC, as Odyssey considers fishing to be hunting. I suppose the best course of action is to, once again, cede authority to the game devs.